Moving view controller to controller directory
